+++ lib/index.d.ts
@@ -1,11 +1,3 @@
-interface WebpackServeOptions {
-    compiler: Object;
-}
-
-interface KoaMiddleware {
-    (context, next) : void
-}
-
 interface Options {
     title?: string;
     theme?: "default" | "dark" | "material";
@@ -13,6 +5,6 @@ interface Options {
     disableWhenValid?: boolean;
 }
 
-declare function webpackServeWaitpage(wsOptions: WebpackServeOptions, options?: Options) : KoaMiddleware;
+declare function webpackDevServerWaitpage(server: Object, options?: Options) : Function;
 
-export = webpackServeWaitpage;
+export = webpackDevServerWaitpage;