+++ test/testMiddleware.js
@@ -2,10 +2,9 @@ const fs = require('fs');
 const path = require('path');
 const ejs = require('ejs');
 const webpack = require('webpack');
-const webpackServe = require('webpack-serve/package.json');
-const SlowDownWebpackPlugin = require('./SlowDownWebpackPlugin');
+const webpackDevServer = require('webpack-dev-server/package.json');
 
-module.exports = (wsOptions, indexFilename, seconds = 2) => {
+module.exports = (server, indexFilename) => {
   const indexFile = path.resolve(__dirname, '..', 'lib', indexFilename);
 
   const state = {
@@ -26,24 +25,17 @@ module.exports = (wsOptions, indexFilename, seconds = 2) => {
     state.progress = (state.progress + 0.1) % 1;
   }, 1000);
 
-  // delay
-  const compilers = wsOptions.compiler.compilers || [wsOptions.compiler];
-  for (let i = 0 ; i < compilers.length ; i++) {
-    const slowDownPlugin = new SlowDownWebpackPlugin(seconds);
-    slowDownPlugin.apply(compilers[i]);
-  }
-
-  return async (ctx/*, next*/) => {
+  return async (req, res/*, next*/) => {
     const total = 200, active = Math.round(total * state.progress);
-    ctx.type = 'html';
-    ctx.body = ejs.render(state.template, {
+    res.setHeader('Content-Type', 'text/html');
+    res.end(ejs.render(state.template, {
       title: 'Development Server (Test)',
       webpackVersion: webpack.version,
-      webpackServeVersion: webpackServe.version,
+      webpackDevServerVersion: webpackDevServer.version,
       progress: [
         [state.progress, 'Building Modules', active + '/200 modules', active + ' active', './' + indexFilename],
         //[(state.progress + 0.1) % 1, 'Building Modules', (active + 1) + '/200 modules', (active + 1) + ' active', './' + indexFilename] // test for multiple compilers
       ]
-    });
+    }));
   }
 };